+++ etna/ensembles/mixins.py
@@ -85,18 +85,7 @@ class SaveEnsembleMixin(SaveMixin):
         self.pipelines: List[BasePipeline]
         self.ts: Optional[TSDataset]
 
-        pipelines = self.pipelines
-        ts = self.ts
-        try:
-            # extract attributes we can't easily save
-            delattr(self, "pipelines")
-            delattr(self, "ts")
-
-            # save the remaining part
-            super().save(path=path)
-        finally:
-            self.pipelines = pipelines
-            self.ts = ts
+        self._save(path=path, skip_attributes=["pipelines", "ts"])
 
         with zipfile.ZipFile(path, "a") as archive:
             with tempfile.TemporaryDirectory() as _temp_dir:
@@ -106,7 +95,7 @@ class SaveEnsembleMixin(SaveMixin):
                 pipelines_dir = temp_dir / "pipelines"
                 pipelines_dir.mkdir()
                 num_digits = 8
-                for i, pipeline in enumerate(pipelines):
+                for i, pipeline in enumerate(self.pipelines):
                     save_name = f"{i:0{num_digits}d}.zip"
                     pipeline_save_path = pipelines_dir / save_name
                     pipeline.save(pipeline_save_path)